+++ NumInWord/Wordable.cs
@@ -2,21 +2,21 @@
 {
     internal abstract class Wordable : IWordable
     {
-        protected int priority;
+        protected short priority;
         
         public int CompareTo(IWordable other)
         {
             return priority.CompareTo(other.GetPriority());
         }
         public abstract string convert();
-        public int GetPriority()
+        public short GetPriority()
         {
             return priority;
         }
-        public abstract bool InsertAnd(int num);
-        public abstract bool IsMatch(int num);
+        public abstract bool InsertAnd(long num);
+        public abstract bool IsMatch(long num);
 
-        public string convert(int num)
+        public string convert(long num)
         {
             string result = string.Empty;
             if (InsertAnd(num))
